pyproject.toml
[project]
name = "openfront-game-bot"
version = "0.1.0"
description = "Basic uv Python project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "discord.py>=2.4.0",
  "python-dotenv>=1.0.1",
]

[tool.uv]
dev-dependencies = [
  "pytest"
]
